% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_function.R
\name{preprocess}
\alias{preprocess}
\title{Function for preprocessing predictor variables for machine learning tasks}
\usage{
preprocess(file, id_var = NULL, target_var = NULL,
  output_file = TRUE, output_removed = FALSE)
}
\arguments{
\item{file}{The file (csv) to handle.}

\item{id_var}{Name of an id variable, if any.}

\item{target_var}{The name of the target variable.}

\item{output_file}{Whether to include an output file.}

\item{output_removed}{Whether to output an additional file detailing which variables where removed and why.}
}
\description{
Function for preprocessing predictor variables for machine learning tasks
}
